Split deployment and adjust config validation
