+++ parallel-install/pkg/config/config_test.go
@@ -16,7 +16,7 @@ func Test_ConfigValidation(t *testing.T) {
 		config = Config{
 			WorkersCount: 0,
 		}
-		err = config.Validate()
+		err = config.ValidateDeletion()
 		assert.Error(t, err)
 		assert.Contains(t, err.Error(), "Workers count cannot be")
 	})
@@ -26,7 +26,7 @@ func Test_ConfigValidation(t *testing.T) {
 			WorkersCount:       1,
 			ComponentsListFile: "/a/file/which/doesnot/exist.json",
 		}
-		err = config.Validate()
+		err = config.ValidateDeletion()
 		assert.Error(t, err)
 		assert.Contains(t, err.Error(), "not found")
 	})
@@ -39,7 +39,7 @@ func Test_ConfigValidation(t *testing.T) {
 			ComponentsListFile: fpath,
 			ResourcePath:       "/a/dir/which/doesnot/exist",
 		}
-		err = config.Validate()
+		err = config.ValidateDeployment()
 		assert.Error(t, err)
 		assert.Contains(t, err.Error(), "not found")
 	})
@@ -53,7 +53,7 @@ func Test_ConfigValidation(t *testing.T) {
 			ResourcePath:       filepath.Dir(fpath),
 			CrdPath:            "/a/dir/which/doesnot/exist",
 		}
-		err = config.Validate()
+		err = config.ValidateDeployment()
 		assert.Error(t, err)
 		assert.Contains(t, err.Error(), "not found")
 	})
@@ -67,7 +67,7 @@ func Test_ConfigValidation(t *testing.T) {
 			ResourcePath:       filepath.Dir(fpath),
 			CrdPath:            filepath.Dir(fpath),
 		}
-		err = config.Validate()
+		err = config.ValidateDeployment()
 		assert.Error(t, err)
 		assert.Contains(t, err.Error(), "Version is empty")
 	})
@@ -82,7 +82,7 @@ func Test_ConfigValidation(t *testing.T) {
 			CrdPath:            filepath.Dir(fpath),
 			Version:            "abc",
 		}
-		err = config.Validate()
+		err = config.ValidateDeployment()
 		assert.NoError(t, err)
 	})
 }
